--- src/obfuscator/namegenerators.lua
@@ -1,6 +1,0 @@
-return {
-	Mangled = require("obfuscator.namegenerators.mangled");
-	MangledShuffled = require("obfuscator.namegenerators.mangled_shuffled");
-	Il = require("obfuscator.namegenerators.Il");
-	Number = require("obfuscator.namegenerators.number");
-}
