+++ src/prometheus/compiler_secure/bytecode.lua
@@ -4,18 +4,18 @@
 -- This Script contains the Bytecode library for Compiling to Prometheus Bytecode
 
 -- For Lua5.1 Compatibility
-local bit32 = require("bit") or require("obfuscator.bit").bit32;
+local bit32 = require("bit") or require("prometheus.bit").bit32;
 local logger = require("logger");
-local util = require("obfuscator.util");
-local vmstrings = require("obfuscator.compiler_secure.vmstrings");
-local Parser = require("obfuscator.parser");
-local enums = require("obfuscator.enums");
+local util = require("prometheus.util");
+local vmstrings = require("prometheus.compiler_secure.vmstrings");
+local Parser = require("prometheus.parser");
+local enums = require("prometheus.enums");
 
 local LuaVersion = enums.LuaVersion;
 local BIT_MAX_8  = 255;
 local BIT_MAX_16 = 65535;
 
-local InstructionKind = require("obfuscator.compiler_secure.instructionkind");
+local InstructionKind = require("prometheus.compiler_secure.instructionkind");
 
 local Bytecode = {
 	InstructionKind = InstructionKind
