+++ src/prometheus/compiler_secure/instructionkind.lua
