+++ src/prometheus/namegenerators.lua
@@ -1,0 +1,6 @@
+return {
+	Mangled = require("prometheus.namegenerators.mangled");
+	MangledShuffled = require("prometheus.namegenerators.mangled_shuffled");
+	Il = require("prometheus.namegenerators.Il");
+	Number = require("prometheus.namegenerators.number");
+}
