+++ src/prometheus/namegenerators/mangled.lua
@@ -1,0 +1,26 @@
+-- This Script is Part of the Prometheus Obfuscator by Levno_710
+--
+-- namegenerators/mangled.lua
+--
+-- This Script provides a function for generation of mangled names
+
+
+local util = require("prometheus.util");
+local chararray = util.chararray;
+
+local idGen = 0
+local VarDigits = chararray("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_");
+local VarStartDigits = chararray("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ");
+
+return function(id, scope)
+	local name = ''
+	local d = id % #VarStartDigits
+	id = (id - d) / #VarStartDigits
+	name = name..VarStartDigits[d+1]
+	while id > 0 do
+		local d = id % #VarDigits
+		id = (id - d) / #VarDigits
+		name = name..VarDigits[d+1]
+	end
+	return name
+end
