+++ src/prometheus/namegenerators/mangled_shuffled.lua
@@ -1,18 +1,17 @@
 -- This Script is Part of the Prometheus Obfuscator by Levno_710
 --
--- namegenerators/mangled.lua
+-- namegenerators/mangled_shuffled.lua
 --
--- This Script provides a function for generation of mangled names
+-- This Script provides a function for generation of mangled names with shuffled character order
 
 
-local util = require("obfuscator.util");
+local util = require("prometheus.util");
 local chararray = util.chararray;
 
-local idGen = 0
 local VarDigits = chararray("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_");
 local VarStartDigits = chararray("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ");
 
-return function(id, scope)
+local function generateName(id, scope)
 	local name = ''
 	local d = id % #VarStartDigits
 	id = (id - d) / #VarStartDigits
@@ -24,3 +23,13 @@ return function(id, scope)
 	end
 	return name
 end
+
+local function prepare(ast)
+	util.shuffle(VarDigits);
+	util.shuffle(VarStartDigits);
+end
+
+return {
+	generateName = generateName, 
+	prepare = prepare
+};
