+++ src/prometheus/namegenerators/number.lua
