+++ test.obfuscated.lua
@@ -1,0 +1,1 @@
+local B={"Go to the store and buy some more!","1 bottle of beer","__call"," bottles of beer","__index",",","No more bottles of beer"," on the wall,\n","","Take one down and pass it around,","drink"," on the wall,","\n","__newindex"}do local n,y=1,14 while n<y do B[n],B[y]=B[y],B[n]n,y=n+1,y-1 end n,y=1,5 while n<y do B[n],B[y]=B[y],B[n]n,y=n+1,y-1 end n,y=6,14 while n<y do B[n],B[y]=B[y],B[n]n,y=n+1,y-1 end end local function o(n)return B[n-39551]end local n=o(39552)local y=o(39563)local g=o(39558)local J=setmetatable({0,[o(39553)]={n..(o(39555)..(y..o(39554))),o(39557)}},{[o(39561)]=function(y,J)y[J]=nil print(y[J]..o(39554))print(y[J]..o(39562))print(n)y[J]=y[J-1]return g..o(39564)end,[o(39556)]=function(n,y,g)rawset(n,y,g or y..o(39560))if rawget(n,y+1)then print(n[y]..o(39564))end end,[o(39559)]=function()return ipairs({g,y})end})print(J[99])for n,y in J()do print(y..o(39554))print(y..o(39562))print(J[o(39553)][n])print(o(39565))end
